test(TreeSelect): add demo test
