+++ example/app.json
@@ -139,6 +139,7 @@
     "van-toast-demo": "./dist/toast/demo/index",
     "van-transition": "./dist/transition/index",
     "van-tree-select": "./dist/tree-select/index",
+    "van-tree-select-demo": "./dist/tree-select/demo/index",
     "van-datetime-picker": "./dist/datetime-picker/index",
     "van-rate": "./dist/rate/index",
     "van-rate-demo": "./dist/rate/demo/index",
