--- example/pages/tree-select/config.js
@@ -1,52 +1,0 @@
-export default {
-  pro1Name: '浙江',
-  pro1: [
-    {
-      text: '杭州',
-      id: 1,
-    },
-    {
-      text: '温州',
-      id: 2,
-    },
-    {
-      text: '宁波',
-      id: 3,
-      disabled: true,
-    },
-    {
-      text: '义乌',
-      id: 4,
-    },
-  ],
-  pro2Name: '江苏',
-  pro2: [
-    {
-      text: '南京',
-      id: 5,
-    },
-    {
-      text: '无锡',
-      id: 6,
-    },
-    {
-      text: '徐州',
-      id: 7,
-    },
-    {
-      text: '苏州',
-      id: 8,
-    },
-  ],
-  pro3Name: '福建',
-  pro3: [
-    {
-      text: '泉州',
-      id: 9,
-    },
-    {
-      text: '厦门',
-      id: 10,
-    },
-  ],
-};
