--- example/pages/tree-select/index.wxss
@@ -1,1 +1,0 @@
-/* pages/tree-select/index.wxss */
