+++ packages/tree-select/demo/index.json
@@ -1,0 +1,8 @@
+{
+  "component": true,
+  "usingComponents": {
+    "van-image": "../../image/index",
+    "van-tree-select": "../../tree-select/index",
+    "demo-block": "../../../example/components/demo-block/index"
+  }
+}
